+++ src/art_profile_builder/lib/APBConfig.hh
@@ -3,6 +3,8 @@
 #include <cstdlib>
 #include <string>
 
+#include <htslib/hts.h>
+
 namespace labw::art_modern {
 /**
  * Art Profile Builder Configuration
@@ -16,6 +18,7 @@ struct APBConfig {
     const bool is_ob;
     const std::string output_1_file_path;
     const std::string output_2_file_path;
+    const htsExactFormat input_file_format;
 };
 
 } // namespace labw::art_modern