+++ docs/News.md
@@ -8,6 +8,7 @@
   - More random numbers are generated in bulk to reduce overhead.
 - Makefile integration test target `testbuild` and `testbuild-mpi` reimplemented in Python to make them run faster.
 - `art_profile_builder` would now raise an error if the input SAM/BAM/FASTQ files are malformed.
+- Memory performance of `stream` FASTA parser largely improved.
 - Miscellaneous bug fixes.
 
 ## 1.2.0 (2025/10/21)